refactor : 신고 항목 표시 유틸 공통화 (#221)
